+++ app/models/folio/atom.rb
@@ -98,6 +98,7 @@ module Folio::Atom
         hint: I18n.t("simple_form.hints.#{klass.name.underscore}.base", default: nil).try(:html_safe),
         structure:,
         editable: klass.editable_in_console?,
+        insertable: klass.insertable_in_console?(site:),
         form_layout: klass::FORM_LAYOUT,
         title: klass.model_name.human,
         molecule: klass.molecule_component_class || klass.molecule_cell_name,