+++ react/src/components/AtomForm/Fields.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import Field from './Field'
 
-export default function Fields ({ atom, index, onChange, onValueChange, startSplittingAtom }) {
+export default function Fields ({ atom, index, onChange, onValueChange, rootAtoms, startSplittingAtom }) {
   return (
     <React.Fragment>
       {Object.keys(atom.record.meta.structure).map((field) => (
@@ -13,6 +13,7 @@ export default function Fields ({ atom, index, onChange, onValueChange, startSpl
           index={index}
           onChange={onChange}
           onValueChange={onValueChange}
+          rootAtoms={rootAtoms}
           startSplittingAtom={startSplittingAtom}
         />
       ))}