+++ react/src/containers/Atoms/tests/Atoms.test.js
@@ -9,6 +9,7 @@ import { SerializedAtom } from 'components/SerializedAtoms'
 import atomsReducer, {
   atomsSelector,
   atomTypesSelector,
+  editAtoms,
   initialState,
   setAtomsData
 } from 'ducks/atoms'
@@ -40,5 +41,27 @@ describe('Atoms', () => {
     const component = renderer.create(<Atoms {...props} />)
     const atomForms = component.root.findAllByType(SerializedAtom)
     expect(atomForms.length).toEqual(2 * 6)
+  })
+
+  it('does not offer non-insertable atoms unless one is being edited', () => {
+    const data = {
+      ...SINGLE_LOCALE_ATOMS,
+      structures: {
+        ...SINGLE_LOCALE_ATOMS.structures,
+        'Folio::Atom::Text': {
+          ...SINGLE_LOCALE_ATOMS.structures['Folio::Atom::Text'],
+          insertable: false
+        }
+      }
+    }
+    let atomsState = atomsReducer(initialState, setAtomsData(data))
+    let state = { atoms: atomsState }
+
+    expect(atomTypesSelector(state).find(({ key }) => key === 'Folio::Atom::Text')).toBeUndefined()
+
+    atomsState = atomsReducer(atomsState, editAtoms('atoms', [0]))
+    state = { atoms: atomsState }
+
+    expect(atomTypesSelector(state).find(({ key }) => key === 'Folio::Atom::Text')).toBeTruthy()
   })
 })