+++ test/cells/folio/console/atoms/previews_cell_test.rb
@@ -3,6 +3,12 @@
 require "test_helper"
 
 class Folio::Console::Atoms::PreviewsCellTest < Folio::Console::CellTest
+  class NonInsertableAtom < Folio::Atom::Base
+    def self.insertable_in_console?(site:)
+      false
+    end
+  end
+
   test "show" do
     create_and_host_site
 
@@ -11,5 +17,6 @@ class Folio::Console::Atoms::PreviewsCellTest < Folio::Console::CellTest
 
     html = cell("folio/console/atoms/previews", { cs: [] }, klass: Folio::Page).(:show)
     assert html.has_css?(".f-c-atoms-previews")
+    assert_not html.has_css?("[data-type='#{NonInsertableAtom.name}']")
   end
 end